% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSUF.R
\name{getSUF}
\alias{getSUF}
\title{Get the hydraulic properties of a given root system. Needs the root architecture from CRootBox, hydraulic properties and soil conditions.}
\usage{
getSUF(table_data = NULL, table_cond = NULL, table_soil = NULL,
  hetero = TRUE, Psi_collar = -15000)
}
\arguments{
\item{table_data}{A data frame with the CRootBox simulation results}

\item{table_cond}{A data frame with the plant conductivity parameters}

\item{table_soil}{A data frame with the soil humidity profile}

\item{hetero}{Do we need to compute the uptake in heterogeneous soil?}
}
\description{
Get the hydraulic properties of a given root system. Needs the root architecture from CRootBox, hydraulic properties and soil conditions.
}
\keyword{root,}
\keyword{water}
