% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.bigmemory.R
\name{marshal.big.matrix}
\alias{marshal.big.matrix}
\alias{marshal.bigmemory}
\alias{marshallable.big.memory}
\title{Marshalling of 'bigmemory' objects}
\usage{
\method{marshal}{big.matrix}(x, ...)

\method{marshallable}{big.memory}(...)
}
\arguments{
\item{x}{A \link[bigmemory:big.matrix]{bigmemory::big.matrix} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'bigmemory' objects
}
\details{
\code{\link[bigmemory:write.big.matrix]{bigmemory::write.big.matrix()}} is used to produce a marshalled version
of the original object.
\code{\link[bigmemory:read.big.matrix]{bigmemory::read.big.matrix()}} is used to reconstruct a version of the
original object from the marshalled object.
}
\examples{
if (requireNamespace("bigmemory", quietly = TRUE)) {
  library(bigmemory)
  
  x <- big.matrix(nrow = 3, ncol = 2, type = "double")
  x[] <- seq_along(x)
  
  ## Marshal
  x_ <- marshal(x)

  ## Unmarshal
  x2 <- unmarshal(x_)

  stopifnot(identical(x2[], x[]))
}

}
