% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.data.table.R
\name{marshal.data.table}
\alias{marshal.data.table}
\title{Marshal and Unmarshal a 'data.table' object}
\usage{
\method{marshal}{data.table}(x, ...)
}
\arguments{
\item{x}{A \link[data.table:data.table]{data.table::data.table} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshal and Unmarshal a 'data.table' object
}
\examples{
if (requireNamespace("data.table", quietly = TRUE)) {
  dt <- data.table::data.table(a = 1:3, b = letters[1:3])

  ## Marshal
  dt_ <- marshal(dt)
  
  ## Unmarshal
  dt2 <- unmarshal(dt_)
  
  stopifnot(identical(dt2, dt))

  ## Marshal and unmarshal again
  dt2_ <- marshal(dt2)
  dt3 <- unmarshal(dt2_)
  
  stopifnot(
    identical(dt2_, dt_),
    identical(dt3, dt2)
  )
}

}
