% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.DBI.R
\name{marshal.DBIConnection}
\alias{marshal.DBIConnection}
\title{Marshalling of 'DBI:DBIConnection' objects (not supported)}
\usage{
\method{marshal}{DBIConnection}(x, ...)
}
\arguments{
\item{x}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
