% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.default.R
\name{marshal.default}
\alias{marshal.default}
\title{Marshalling of an R objects}
\usage{
\method{marshal}{default}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of an R objects
}
\details{
This is the default, fallback method for marshalling of an
\R object.  It assumes that the object do not need specific
marshalling, but can be serialized as-is.  All this default
method does is wrapping the original object up in a
\code{marshalled} object, so that the original object can be
recovered via \code{unmarshalled()}.
}
\examples{
x <- data.frame(a = 1:3, b = letters[1:3])
x_ <- marshal(x)
x2 <- unmarshal(x_)
stopifnot(identical(x2, x))


}
