% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.h2o.R
\name{marshal.H2OAutoML}
\alias{marshal.H2OAutoML}
\alias{marshal.H2OMultinomialModel}
\alias{marshal.H2OBinomialModel}
\alias{marshal.H2ORegressionModel}
\alias{marshallable.H2OAutoML}
\alias{marshallable.H2OMultinomialModel}
\alias{marshallable.H2OBinomialModel}
\alias{marshallable.H2ORegressionModel}
\title{Marshalling of 'h2o' objects}
\usage{
\method{marshal}{H2OAutoML}(x, ...)

\method{marshal}{H2OMultinomialModel}(x, ...)

\method{marshal}{H2OBinomialModel}(x, ...)

\method{marshal}{H2ORegressionModel}(x, ...)

\method{marshallable}{H2OAutoML}(...)

\method{marshallable}{H2OMultinomialModel}(...)

\method{marshallable}{H2OBinomialModel}(...)

\method{marshallable}{H2ORegressionModel}(...)
}
\arguments{
\item{x}{An "h2o" object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'h2o' objects
}
\details{
\code{\link[h2o:h2o.save_mojo]{h2o::h2o.save_mojo()}} and \code{\link[h2o:h2o.saveModel]{h2o::h2o.saveModel()}} are used to produce
a marshalled version of the original object.
\code{\link[h2o:h2o.import_mojo]{h2o::h2o.import_mojo()}} and \code{\link[h2o:h2o.loadModel]{h2o::h2o.loadModel()}} are used to
reconstruct a version of the original object from the marshalled object.
}
