% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.magick.R
\name{marshal.magick-image}
\alias{marshal.magick-image}
\alias{marshallable.magick-image}
\title{Marshalling of 'magick' objects (not supported)}
\usage{
\method{marshal}{`magick-image`}(x, ...)

\method{marshallable}{`magick-image`}(...)
}
\arguments{
\item{x}{A \link[magick:magick]{magick:magick-image} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
