% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.ncdf4.R
\name{marshal.ncdf4}
\alias{marshal.ncdf4}
\title{Marshal and Unmarshal an 'ncdf4' object}
\usage{
\method{marshal}{ncdf4}(x, ...)
}
\arguments{
\item{x}{A \link[ncdf4:nc_create]{ncdf4:ncdf4} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshal and Unmarshal an 'ncdf4' object
}
\details{
\code{\link[base:readBin]{base::readBin()}} is used to produce a marshalled version
of the original object.
\code{\link[base:readBin]{base::writeBin()}} and \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}} are used to
reconstruct a version of the original object from the
marshalled object.
}
\examples{
if (requireNamespace("ncdf4", quietly = TRUE)) {
  library(ncdf4)
  
  ## Create a temporary NCDF file
  tf <- tempfile(fileext = ".nc")
  
  x <- ncvar_def("x", units = "count", dim = list())
  nf <- nc_create(tf, x)
  ncvar_put(nf, x, 42)
  print(nf)
  nc_close(nf)
  rm(list = c("nf", "x"))

  ## Open NCDF file
  nf <- nc_open(tf)
  y <- ncvar_get(nf, "x")

  ## Marshal
  nf_ <- marshal(nf)

  ## Unmarshal
  nf2 <- unmarshal(nf_)

  y2 <- ncvar_get(nf2, "x")
  stopifnot(identical(y2, y))
}
}
