% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.parallel.R
\name{marshal.SOCKcluster}
\alias{marshal.SOCKcluster}
\alias{marshal.SOCKnode}
\alias{marshal.SOCK0node}
\title{Marshal and Unmarshal a 'parallel' cluster object (not supported)}
\usage{
\method{marshal}{SOCKcluster}(x, ...)
}
\arguments{
\item{x}{A \link[magick:magick]{magick:magick-image} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
