% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.parsnip.R
\name{marshal.model_fit}
\alias{marshal.model_fit}
\title{Marshalling of 'parsnip' objects}
\usage{
\method{marshal}{model_fit}(x, ...)
}
\arguments{
\item{x}{A \code{parnsip:model_fit} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'parsnip' objects
}
\details{
The \code{fit} element of the \code{model_fit} object is marshalled.
}
\examples{
if (requireNamespace("parsnip", quietly = TRUE) && requireNamespace("xgboost", quietly = TRUE)) {
  library(parsnip)
  
  ## Adopted from example("boost_tree", package = "parsnip")
  model <- boost_tree(mode = "classification", trees = 20L, engine = "xgboost")  
  model <- set_mode(model, "regression")
  fit <- fit(model, mpg ~ ., data = datasets::mtcars)
  
  ## Marshal
  fit_ <- marshal(fit)

  ## Unmarshal
  fit2 <- unmarshal(fit_)

  ## Marshal again
  fit2_ <- marshal(fit2)

  ## Assert identity
  stopifnot(
    all.equal(fit2_, fit_)
  )
  
  fit3 <- unmarshal(fit2_)

  ## Assert identity
  stopifnot(
    all.equal(fit3, fit2)
  )
}  
}
