% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.raster.R
\name{marshal.RasterLayer}
\alias{marshal.RasterLayer}
\alias{marshallable.RasterLayer}
\title{Marshalling of 'raster' objects}
\usage{
\method{marshal}{RasterLayer}(x, ...)

\method{marshallable}{RasterLayer}(x, ...)
}
\arguments{
\item{x}{A \link[raster:raster]{raster:RasterLayer} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'raster' objects
}
\details{
\code{\link[raster:writeRaster]{raster::writeRaster()}} is used to produce a marshalled version
of the original object.
\code{\link[raster:raster]{raster::raster()}} is used to reconstruct a version of the
original object from the marshalled object.
}
\section{WARNING - Numerical identical results cannot be guaranteed}{

Marshalling of \code{RasterLayer} objects is \emph{leaky}.  More precisely,
the method \emph{cannot} guarantee that multiple rounds of marshalling
and unmarshalling produce numerically identical objects.
For further information, see (1).
}

\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  r <- raster::raster(system.file("external/test.grd", package = "raster"))
  print(r)
  
  ## Marshal
  r_ <- marshal(r)

  ## Unmarshal
  r2 <- unmarshal(r_)
  print(r2)
}
}
\references{
\enumerate{
\item Issue #311 - 'Incorrect large values saved to disk with writeRaster()',
\strong{raster} GitHub repostory, 2023-05-17,
\url{https://github.com/rspatial/raster/issues/311}.
}
}
