% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.reticulate.R
\name{marshal.python.builtin.object}
\alias{marshal.python.builtin.object}
\alias{marshallable.python.builtin.object}
\title{Marshalling of 'reticulate' objects (not supported)}
\usage{
\method{marshal}{python.builtin.object}(x, ...)

\method{marshallable}{python.builtin.object}(...)
}
\arguments{
\item{x}{"reticulate" object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
