% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.rJava.R
\name{marshal.jclassName}
\alias{marshal.jclassName}
\alias{marshal.jobjRef}
\alias{marshallable.jobjRef}
\alias{marshallable.jclassName}
\title{Marshalling of 'rJava' objects (not supported)}
\usage{
\method{marshal}{jclassName}(x, ...)

\method{marshal}{jobjRef}(x, ...)

\method{marshallable}{jclassName}(...)

\method{marshallable}{jobjRef}(...)
}
\arguments{
\item{x}{"rJava" object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
