% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.sparklyr.R
\name{marshal.tbl_spark}
\alias{marshal.tbl_spark}
\alias{marshallable.tbl_spark}
\title{Marshalling of 'sparklyr' objects (not supported)}
\usage{
\method{marshal}{tbl_spark}(x, ...)

\method{marshallable}{tbl_spark}(...)
}
\arguments{
\item{x}{"sparklyr" object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
