% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.terra.R
\name{marshal.SpatVector}
\alias{marshal.SpatVector}
\alias{marshal.terra}
\title{Marshal and Unmarshal an 'terra' object}
\usage{
\method{marshal}{SpatVector}(terra, ...)
}
\arguments{
\item{terra}{An \link[terra:SpatVector-class]{terra::SpatVector}.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshal and Unmarshal an 'terra' object
}
\examples{
if (requireNamespace("terra", quietly = TRUE)) {
  file <- system.file("ex/lux.shp", package = "terra")
  v <- terra::vect(file)
  
  ## Marshal SpatVector object
  v_ <- marshal(v)
  
  ## Unmarshal SpatVector object
  v2 <- unmarshal(v_)
  
  stopifnot(all.equal(v2, v, check.attributes = FALSE))
}

}
