% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.train.R
\name{marshal.train}
\alias{marshal.train}
\title{Marshal and Unmarshal a 'train' object}
\usage{
\method{marshal}{train}(train, ...)
}
\arguments{
\item{train}{A \link[caret:train]{caret::train} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshal and Unmarshal a 'train' object
}
\details{
\code{\link[bundle:bundle]{bundle::bundle()}} is used to produce a marshalled version
of the original object.
\code{\link[bundle:bundle]{bundle::unbundle()}} is used to reconstruct a version of the
original object from the marshalled object.
}
