% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.udpipe.R
\name{marshal.udpipe_model}
\alias{marshal.udpipe_model}
\alias{marshallable.udpipe_model}
\title{Marshalling of 'udpipe' objects (not supported)}
\usage{
\method{marshal}{udpipe_model}(x, ...)

\method{marshallable}{udpipe_model}(...)
}
\arguments{
\item{x}{"udpipe" object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
\emph{Warning: Objects of this class are not possible to marshal.
If attempted, an error is produced.}
}
