% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.xgb.DMatrix.R
\name{marshal.xgb.DMatrix}
\alias{marshal.xgb.DMatrix}
\title{Marshal and Unmarshal an 'xgboost::xgb.DMatrix' object}
\usage{
\method{marshal}{xgb.DMatrix}(x, ...)
}
\arguments{
\item{x}{An \link[xgboost:xgb.DMatrix]{xgboost::xgb.DMatrix} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshal and Unmarshal an 'xgboost::xgb.DMatrix' object
}
\details{
\code{\link[xgboost:xgb.DMatrix.save]{xgboost::xgb.DMatrix.save()}} is used to produce a marshalled version
of the original object.
\code{\link[xgboost:xgb.DMatrix]{xgboost::xgb.DMatrix()}} is used to reconstruct a version of the
original object from the marshalled object.
}
