% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.xgboost.R
\name{marshal.xgb.DMatrix}
\alias{marshal.xgb.DMatrix}
\alias{marshal.xgb.Booster}
\alias{marshallable.xgb.Booster}
\alias{marshallable.xgb.DMatrix}
\title{Marshalling of 'xgboost' objects}
\usage{
\method{marshal}{xgb.DMatrix}(x, ...)

\method{marshal}{xgb.Booster}(x, ...)

\method{marshallable}{xgb.Booster}(...)

\method{marshallable}{xgb.DMatrix}(...)
}
\arguments{
\item{x}{An \link[xgboost:xgb.DMatrix]{xgboost::xgb.DMatrix} or an \code{xgboost:xgb.Booster} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'xgboost' objects
}
\details{
\code{\link[xgboost:xgb.DMatrix.save]{xgboost::xgb.DMatrix.save()}} is used to produce a marshalled version
of the original object.
\code{\link[xgboost:xgb.DMatrix]{xgboost::xgb.DMatrix()}} is used to reconstruct a version of the
original object from the marshalled object.
}
