% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.xml2.R
\name{marshal.xml_document}
\alias{marshal.xml_document}
\alias{marshal.xml2}
\alias{marshal.xml_nodeset}
\alias{marshallable.xml_document}
\alias{marshallable.xml_nodeset}
\title{Marshalling of 'xml2' objects}
\usage{
\method{marshal}{xml_document}(xml2, ...)

\method{marshal}{xml_nodeset}(xml2, ...)

\method{marshallable}{xml_document}(...)

\method{marshallable}{xml_nodeset}(...)
}
\arguments{
\item{xml2}{A \link[xml2:xml_document-class]{xml2::xml_document} or similar.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'xml2' objects
}
\details{
\code{\link[xml2:xml_serialize]{xml2::xml_serialize()}} is used to produce a marshalled version
of the original object.
\code{\link[xml2:xml_serialize]{xml2::xml_unserialize()}} is used to reconstruct a version of the
original object from the marshalled object.
}
\examples{
if (requireNamespace("xml2", quietly = TRUE)) {
  doc <- xml2::read_xml("<body></body>")

  ## Marshal 'xml_document' object
  doc_ <- marshal(doc)

  ## Unmarshal 'xml_document' object
  doc2 <- unmarshal(doc_)

  stopifnot(all.equal(doc2, doc))
}
}
