% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshallable.R
\name{marshallable}
\alias{marshallable}
\title{Check if R object can be marshalled}
\usage{
marshallable(...)
}
\arguments{
\item{\dots}{The object to be checked,
followed by additional arguments passed to the specific S3 method.}
}
\value{
\code{TRUE} if the object can be marshalled,
\code{FALSE} if it cannot be marshalled, and
\code{NA} if it is not known whether the object can be marshalled.
}
\description{
Check if R object can be marshalled
}
