% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSUF.R
\name{getSUF}
\alias{getSUF}
\title{MARSHAL: MAize Root System Hydraulic Architecture Solver}
\usage{
getSUF(
  table_data = NULL,
  table_cond = NULL,
  table_soil = NULL,
  hetero = TRUE,
  Psi_collar = -15000,
  soil_param = NULL,
  verbatim = FALSE
)
}
\arguments{
\item{table_data}{A data frame with the CRootBox simulation results. The root system architecture should be descretize by small root segment. Needs column node1ID, node2ID, branchID, x1, y1, z1, x2, y2, z2, radius [cm], length [cm], R, G, B, time [d], type, age [d], rep.}

\item{table_cond}{A data frame with the plant conductivity parameters. Needs column id, order_id [root type], order [root type name], type ("Kr" or "Kx"), x (distance from the apex)[cm], y (value for Kx [cm4 hPa-1 d-1] or Kr [cm hPa-1 d-1]).}

\item{table_soil}{A data frame with the soil humidity profile. Needs column: id, z [cm], psi [hPa].}

\item{hetero}{Do we need to compute the uptake in heterogeneous soil? Default: hetero = TRUE}

\item{Psi_collar}{Water potential at the collar (boundary condition). Default: -15.000 [hPa]}

\item{soil_param}{A data frame with the soil hydraulic parameter data(n,alpha,Ksat,lambda,Q_r,Q_s)}

\item{verbatim}{If true, outputs the different step of the procedure}
}
\description{
Get the hydraulic properties of a given root system. Needs the root architecture from CRootBox, hydraulic properties and soil conditions.
}
\author{
Felicien Meunier, Adrien Heymans
}
\keyword{root,}
\keyword{water}
