% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.connection.R
\name{marshal.connection}
\alias{marshal.connection}
\alias{marshallable.connection}
\title{Marshalling of R connections}
\usage{
\method{marshal}{connection}(con, ...)

\method{marshallable}{connection}(con, ...)
}
\arguments{
\item{con}{A \link[base:connection]{connection}.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of R connections
}
\section{Limitations}{

Not all connections can be marshalled, specifically we:
\itemize{
\item cannot marshal connections stdin (0), stdout (1), and stderr (2)
\item can only marshal \emph{read-only} connections
\item can only marshal \emph{unopened} or \emph{seekable} connections
}
}

\examples{
tf <- tempfile()
cat(file = tf, letters, sep = "")

## Read-only connection
con <- file(tf, open = "rb")

bfr <- readChar(con, nchars = 4L)
print(bfr)  ## "abcd"

## Marshal read-only connection, which records the
## current state, including the current file position.
con_ <- marshal(con)

## Unmarshal connection, which restores the state
## of the original connection, including the current
## file position
con2 <- unmarshal(con_)
stopifnot(
  all.equal(summary(con2), summary(con)),
  identical(seek(con2), seek(con))
)

bfr <- readChar(con, nchars = 4L)
print(bfr)
bfr2 <- readChar(con2, nchars = 4L)
print(bfr2)
stopifnot(identical(bfr2, bfr))

## Cleanup
close(con)
close(con2)
file.remove(tf)
con <- url("https://www.r-project.org")
print(con)

## Marshal read-only connection, which records the
## current state, including the current file position.
con_ <- marshal(con)

## Unmarshal connection, which restores the state
## of the original connection
con2 <- unmarshal(con_)
print(con2)

stopifnot(all.equal(summary(con2), summary(con)))

bfr <- readChar(con, nchars = 100L)
print(bfr)
bfr2 <- readChar(con2, nchars = 100L)
print(bfr2)
stopifnot(identical(bfr2, bfr))

## Cleanup
close(con)
close(con2)

}
