% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{marshal}
\alias{marshal}
\alias{unmarshal}
\title{Marshalling of R objects}
\usage{
marshal(...)

unmarshal(...)
}
\arguments{
\item{\dots}{The object to be marshalled, or unmarshalled,
followed by additional arguments passed to the specific S3 method.}
}
\value{
\code{marshal()} returns a \code{marshalled} object, which is a list with
components:
\itemize{
\item \code{marshalled}: marshalled version of the original object
\item \code{unmarshal}: function that takes the \code{marshalled} object as
input and returns an unmarshalled version of the original object.
}

\code{unmarshal()} returns an unmarshaled version of the original object.
}
\description{
Marshalling of R objects
}
