% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{marshal}
\alias{marshal}
\alias{unmarshal}
\title{Marshal and Unmarshal R Objects}
\usage{
marshal(...)

unmarshal(...)
}
\arguments{
\item{\dots}{The objects to be marshalled, or unmarshalled,
followed by additional arguments passed to the S3 method.}
}
\value{
A marshaled version of the original object.

An unmarshaled version of the original object.
}
\description{
Marshal and Unmarshal R Objects
}
