% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.rstan.R
\name{marshal.stanfit}
\alias{marshal.stanfit}
\title{Marshalling of 'rstan' objects}
\usage{
\method{marshal}{stanfit}(x, ...)
}
\arguments{
\item{x}{A \code{rstan:stanfit} object.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'rstan' objects
}
