% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.XML.R
\name{marshal.XMLAbstractNode}
\alias{marshal.XMLAbstractNode}
\alias{marshal.XML}
\alias{marshal.XMLAbstractDocument}
\title{Marshalling of 'XML' objects}
\usage{
\method{marshal}{XMLAbstractNode}(xml, ...)

\method{marshal}{XMLAbstractDocument}(xml, ...)
}
\arguments{
\item{xml}{An \link[XML:XMLAbstractNode-class]{XML::XMLAbstractNode} or
\link[XML:XMLAbstractDocument-class]{XML::XMLAbstractDocument}.}

\item{\dots}{Not used.}
}
\value{
A \code{marshalled} object as described in \code{\link[=marshal]{marshal()}}.
}
\description{
Marshalling of 'XML' objects
}
\details{
\code{\link[XML:xmlSerializeHook]{XML::xmlSerializeHook()}} is used to produce a marshalled version
of the original object.
\code{\link[XML:xmlSerializeHook]{XML::xmlDeserializeHook()}} is used to reconstruct a version of the
original object from the marshalled object.
}
\examples{
if (requireNamespace("XML", quietly = TRUE)) {
  node <- XML::xmlParseString("<a><b>c</b></a>")
  print(node)
  
  ## Marshal XMLAbstractNode object
  node_ <- marshal(node)
  ## Unmarshal XMLAbstractNode object
  node2 <- unmarshal(node_)
  print(node2)
  
  stopifnot(all.equal(node2, node))
  
  
  doc <- XML::xmlParse(system.file("exampleData", "tides.xml", package = "XML"))
  
  ## Marshal XMLAbstractDocument object
  doc_ <- marshal(doc)
  ## Unmarshal XMLAbstractDocument object
  doc2 <- unmarshal(doc_)
  
  stopifnot(all.equal(doc2, doc))
}
}
